Added error handling for some cases

+++ server/controllers/auth.js
@@ -12,6 +12,8 @@ const sendOtp = async (req, res) => {
   };
   try {
     const otp = generateOtp();
+    if (!req.body.email)
+      return res.status(400).json({ error: "Email cannot be empty" });
 
     await OTP.deleteMany({ email: req.body.email });
 
@@ -27,74 +29,108 @@ const sendOtp = async (req, res) => {
 };
 
 const verifyOtp = async (req, res) => {
-  const { email, otp } = req?.body;
-  const result = await OTP.findOne({ email, otp });
+  try {
+    const { email, otp } = req?.body;
 
-  if (result) {
-    res.status(200).json({
-      success: true,
-      message: "OTP verified successfully",
-    });
-  } else {
-    console.log("OTP entered is wrong");
-    res.status(400).json({ success: false, message: "OTP entered is wrong" });
+    if (!email || !otp)
+      return res.status(400).json({ error: "Email and OTP cannot be empty" });
+    const result = await OTP.findOne({ email, otp });
+
+    if (result) {
+      res.status(200).json({
+        success: true,
+        message: "OTP verified successfully",
+      });
+    } else {
+      console.log("OTP entered is wrong");
+      res.status(400).json({ success: false, message: "OTP entered is wrong" });
+    }
+  } catch (err) {
+    console.log(err.message);
+    res.status(400).json({ success: false, error: err.message });
   }
 };
 
 const registerUser = async (req, res) => {
-  const { email, password } = req.body;
-  console.log({ email, password });
-  const result = await User.findOne({ email: email });
+  try {
+    const { email, password } = req.body;
+    console.log({ email, password });
+    if (!email || !password)
+      return res
+        .status(400)
+        .json({ error: "Email and password cannot be empty" });
+    const result = await User.findOne({ email: email });
 
-  if (!result) {
-    await bcrypt.hash(password, 10).then((hash) => {
-      // const newUser = await User.create(user);
-      User.create({ ...req.body, password: hash })
-        .then((user) => {
-          res.status(200).json({
-            success: true,
-            user: user.email,
-            message: "User added successfully",
+    if (!result) {
+      await bcrypt.hash(password, 10).then((hash) => {
+        // const newUser = await User.create(user);
+        User.create({ ...req.body, password: hash })
+          .then((user) => {
+            res.status(200).json({
+              success: true,
+              user: user.email,
+              message: "User added successfully",
+            });
+          })
+          .catch((err) => {
+            res.status(400).json({
+              success: false,
+              message: err.message,
+            });
           });
-        })
-        .catch((err) => {
-          res.json(400).json({
-            success: false,
-            message: err.message,
-          });
-        });
+      });
+    } else {
+      return res
+        .status(400)
+        .json({ success: false, message: "User already exists" });
+    }
+  } catch (err) {
+    console.log(err);
+    res.status(400).json({
+      success: false,
+      message: err.message,
     });
-  } else {
-    return res
-      .status(400)
-      .json({ success: false, message: "User already exists" });
   }
 };
 
 const loginUser = async (req, res) => {
-  const { email, password } = req?.body;
+  try {
+    const { email, password } = req?.body;
 
-  const user = await User.findOne({ email: email });
-  console.log("User logged in successfully");
-  if (user) {
-    bcrypt.compare(password, user.password, async (err, result) => {
-      if (result) {
-        res.status(200).json({
-          success: true,
-          user: user.email,
-          message: "User logged in successfully",
-        });
-      } else {
-        res.status(400).json({
-          success: false,
-          message: "Invalid credentials",
-        });
-      }
-    });
-  } else {
+    if (!email || !password)
+      return res.json({
+        success: false,
+        message: "Email and password cannot be empty",
+      });
+
+    const user = await User.findOne({ email: email });
+    console.log("User logged in successfully");
+    if (user) {
+      bcrypt.compare(password, user.password, async (err, result) => {
+        if (result) {
+          res.status(200).json({
+            success: true,
+            user: user.email,
+            message: "User logged in successfully",
+          });
+        } else {
+          res.status(400).json({
+            success: false,
+            message: "Invalid credentials",
+          });
+        }
+      });
+    } else {
+      res.status(400).json({
+        success: false,
+        message: "User does not exist",
+      });
+    }
+  } catch (err) {
+    console.log(err);
     res.status(400).json({
       success: false,
-      message: "User does not exist",
+      message: err.message,
     });
   }
 };
